requirements.txt
Flask
mysql-connector-python
requests
python-dotenv

